SB-744: Extract the Maven 2.x layout provider code into a separate module
- Committing work-in-progress.

SB-729: Remove all the test repositories from the strongbox.xml file

- Committing more related fixes (regarding the metadata tests).
